Move model content to new generic.py module
